+++ app/list.tsx
@@ -2,9 +2,7 @@ import React from 'react'
 import { Box } from 'theme-ui'
 import type { Preprints } from '../types/preprint'
 import { formatDate, authorList } from '../utils/formatters'
-import Row from '../components/row'
-import Column from '../components/column'
-import Badge from '../components/badge'
+import { Badge, Column, Row } from '../components'
 
 interface ListViewProps {
   preprints: Preprints
